Fix red light bug and input z axis filtering bug.
    * Make reward_log_sampling, and reset_accel_on_stop optional and turned off by default. (Makes current training run unstable).

   * Make increasing ego box configurable for eval eval_perceived_size_margin_m
    I did not change the default.

    * Add option to add layer norm to the heads as well. Turned off by default

    * Fix bug where the input items are clipped if height is difference is more than 4 meters. This removed nearby cars on steep slopes in town03. I removed this filter because the ego get x,y,z as input so can judge by itself if a car is not relevant.

    * Fix some visu bugs and compiler warnings

    * Only keep the lastest model checkpoint and delete the older once to save disk space.

    * Fix red light bug where the model got red light infractions when inside intersections. Now only infractions are incurred when the model enters an intersection.

+++ pufferlib/ocean/drive/render.h
@@ -283,9 +283,9 @@ Client *make_client(Drive *env) {
             FILE *f = fopen(lock_file, "r");
             if (f) {
                 pid_t pid = -1;
-                fscanf(f, "%d", &pid);
+                int parsed_field_count = fscanf(f, "%d", &pid);
                 fclose(f);
-                if (pid > 0 && kill(pid, 0) != 0) {
+                if (parsed_field_count == 1 && pid > 0 && kill(pid, 0) != 0) {
                     unlink(lock_file);
                     unlink(socket_file);
                 }
